xrd-desktop-cursor: add desktop cursor and use it in overlay_client
Also reworks the xrd_input_synth api a little so xrd_input_synth_move_cursor() takes the intersection point.
This is to reduce confusion, because xrd_overlay_desktop_cursor_update() requires to know the 2d offset on the overlay in world units, while the input synth requires to know it in pixels.
Passing the intersection point to both makes this an implementation detail of each class.