overlay-pointer-tip: Make animation a bit larger and fix overlay size.
It's now easier to increase the texture size with a single value.
The standard background animation will start from filling the entire texture no matter how big it is.

Previously the overlay size was set based on the texture width, which was wrong because increasing the texture size and keeping the overlay size, made the visible size of the pointer smaller.
Now the overlay size takes the texture width into account.