overlay-pointer-tip: Remove "no hmd pose" error message
It's not actually an error, it was more a debug message.
It happens when the HMD is out of basestation range (user takes HMD off and puts it somewhere)
as well as when opening the SteamVR Keyboard.

In those cases the pointer tip size doesn't really matter, so just silently set it to default size.