overlay-window: pass ppm to constructor and fix up buttons a bit
We want to crete windows with an individual initial ppm setting because on a desktop WM windows can be scaled arbitrarily.

Consider a wayland setup with one hidpi and one low dpi monitor.
On the hidpi monitor the same window might be twice the size as on the hidpi monitor as when moved to the low dpi monitor but in XR we still want them to be the same.
The desktop WM would have to update the ppm of a window in this case.