window-manager: Rework head following overlay
Instead of intersecting lines from windows to the center of the view, use angles to describe the window's position.
This fixes some issues with the window being behind the user and the intersection going through the view plane first.

It also makes consistent velocity calculations much easier since the velocity is now expressed in terms of angles, not sphere sections.

One aspect is worse though: The snapping and window movement is perpendicular to the FOV edges now.
This will be solved in a follow up commit.