client: Keep controllers in dynamic list
Instead of the [0,1] ids for left and right, use a guint64 handle per controller.
In the case of OpenVR, OpenVR's tracked device index is used but it could be any other unique mapping.

Slightly ugly workaround:
There is no activate event for controllers already active when xrdesktop starts.
These controllers will be "activated" when the first pose action arrives.