window: Implement selection for overlay window, stub for scene window
Selection is a purely visual effect on the window.
Selected windows are marked blue, deselected windows are marked gray.
When the selection is ended, the window is unmarked.

For overlay windows, the only viable method for marking overlays is by applying a color tint.
For scene windows a nicer effect will be implemented.