client: Add overlay<->scene switch functionality
As a first implementation we destroy all XrdWindows and create new ones, which requires the desktop to actively update its XrdWindow references.
In a later step, we will move the backend independent stuff to a struct, which will be what desktops will use to hold references to vr windows. These structs will persist over the switch.

Note:
Restoring containers is not yet implemented.
The window control button container will be destroyed on the switch and will have to be summoned again with the menu/b button.